backend/requirements.txt
Flask
flask-socketio
flask-cors
pymongo
geopy
requests
python-dotenv
eventlet
